requirements.txt
arize-phoenix[evals]==4.7.1
datasets==2.20.0
gcsfs==2024.6.1
langchain==0.1.20
llama_index==0.10.59
matplotlib==3.9.1
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.37.1
openinference-instrumentation-langchain==0.1.25
openinference-instrumentation-llama-index==2.2.0
opentelemetry-exporter-otlp==1.26.0 
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
pandas==2.2.2
ragas==0.1.4
seaborn==0.13.2
trulens_eval==0.33.0